Lab02S01/requirements.txt
requests
pandas
numpy
scipy
matplotlib
tqdm
python-dotenv
